Split components into several for re-use
